+++ tests/testthat/test_fluxEstimates.R
@@ -34,6 +34,7 @@ test_that("detect special case of numericall equal fluxes",{
   # https://github.com/bgctw/RespChamberProc/issues/2
   timesR <- 0:81
   concR <- rep(88.01, length(timesR))
+  concR[3] <- NA    # test dealing with NA
   #plot( concR ~ timesR )
   ds <- data.frame(
     CO2_dry = concR, TIMESTAMP = timesR, TA_Avg = 20, Pa = 101*1000)
@@ -44,3 +45,15 @@ test_that("detect special case of numericall equal fluxes",{
   expect_true( is.na(res$flux) )
 })			
 			
+test_that("detect special case no finite concentrations",{
+  timesR <- 0:81
+  concR <- rep(NA_real_, length(timesR))
+  #plot( concR ~ timesR )
+  ds <- data.frame(
+    CO2_dry = concR, TIMESTAMP = timesR, TA_Avg = 20, Pa = 101*1000)
+  #trace(calcClosedChamberFlux, recover)		#untrace(calcClosedChamberFlux)
+  expect_warning(
+    res <- calcClosedChamberFlux(ds)
+  )
+  expect_true( is.na(res$flux) )
+})